ecds: move filter config files out of http/
