+++ source/extensions/filters/http/health_check/BUILD
@@ -39,7 +39,7 @@ envoy_cc_extension(
     hdrs = ["config.h"],
     # Legacy test use. TODO(#9953) clean up.
     extra_visibility = [
-        "//test/common/filter/http:__subpackages__",
+        "//test/common/filter:__subpackages__",
         "//test/integration:__subpackages__",
         "//test/server:__subpackages__",
     ],
