+++ test/common/filter/BUILD
@@ -9,11 +9,11 @@ licenses(["notice"])  # Apache 2
 envoy_package()
 
 envoy_cc_test(
-    name = "filter_config_discovery_impl_test",
-    srcs = ["filter_config_discovery_impl_test.cc"],
+    name = "config_discovery_impl_test",
+    srcs = ["config_discovery_impl_test.cc"],
     deps = [
         "//source/common/config:utility_lib",
-        "//source/common/filter/http:filter_config_discovery_lib",
+        "//source/common/filter:config_discovery_lib",
         "//source/common/json:json_loader_lib",
         "//source/extensions/filters/http/health_check:config",
         "//source/extensions/filters/http/router:config",
