+++ test/extensions/filters/network/http_connection_manager/config_test_base.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "source/common/filter/http/filter_config_discovery_impl.h"
+#include "source/common/filter/config_discovery_impl.h"
 #include "source/common/http/date_provider_impl.h"
 #include "source/common/network/address_impl.h"
 #include "source/extensions/filters/http/common/factory_base.h"
